World Layoffs Data Set
This .sql file contains Data Cleaning queries for the Layoffs Data set using MySQL.

It is a guided project and I learned a lot of new syntax. Also, since it is a little bit advanced, I learned applying almost all the basics such as Create Table, staging of the raw data for a more professional approach, Updating the working table, and details of deleting rows and/or columns.

I get a chance to combine the basics of different statements to properly Remove duplicates, Standardize the data, work with Nulls, and carefully delete row/s and/or columns. This Data Cleaning prepares the table for Exploratory Data Analysis (EDA).